Use start+while pattern for directives
The directives appear to not yield their scope at the end of some
blocks, particularly when followed by list entries. Convert to using a
start+while pattern instead, checking that each line in the block
matches the indentation, rather than checking for lines that don't
match.

Puts the extra indentation in a positive lookahead to avoid consuming
the character, as this would break nested blocks that rely on the
opening whitespace to find the next indentation level.

Fixes #13.